backend/requirements.txt
# Core dependencies
python-dotenv==1.1.0
flask
flask-cors
requests

# Google AI & ADK
google-genai==1.46.0
google-generativeai==0.8.5
google-adk

# A2A Protocol
a2a-sdk

# Salesforce
simple-salesforce==1.12.6

# Database & Storage  
snowflake-connector-python
boto3

# Document processing
pdfplumber
Pillow

# Web server for A2A agents
uvicorn
starlette
httpx

# Other (from original)
yfinance==0.2.56
psutil==6.1.1
litellm==1.66.3